add MedalSummaryTopView
